Improve docker experience on CLI
Wraps `build`, `run`, `attach` and `kill` tasks in individual shell scripts.

Also, dockerfile is modified to ensure attaching a running container will
have the environment setup for ros.

The development process requires execution of specific set of commands to
execute most of the tasks above. That is reducing time efficiency to do an
iteration on CLI.

This change aims to reduce amount of commands required to perform above
tasks.