Implement waypoint updater
Updates the waypoints to follow along the lane at 50Hz with respect to environment (e.g., traffic).

Compared to walkthrough boilerplate code in the lecture, we applied 2 fundamental differences
both of which contributes to the ros node performance.

1. The waypoint horizon is set as 50 points that seems to be good enough for making smooth decisions.
2. The slow down waypoint horizon is limited to the stop destination. The code avoids further waypoints
that is within the horizon

The slow down logic has been kept the same as the walkthrough boilerplate code after trying following
Ease-in & Ease-out alternative. That logic seems to face issues with controller nodes in latter stages of
slowing down

```python
class EaseInEaseOutInterpolator(object):
  """
  Source: https://math.stackexchange.com/a/121755
  """

  @staticmethod
  def interpolate(x):
    """
    Interpolates the x within 0.0 and 1.0 bounds to a ease-in and ease-out
    trajectory

    Everything outside the expected [0, 1] range will be rounded to the closest bound
    """
    if x > 1.0:
      return 1.0
    elif x < 0.0:
      return 0.0

    power = 1.5
    x_power = x**power
    return (x_power)/(x_power + (1-x)**power)
```

that could be used as

```python
vel = 0.0

if i <= stop_idx:
    vel = EaseInEaseOutInterpolator.interpolate((stop_idx - i) / LOOKAHEAD_WPS_FLOAT) * ref_velocity
```